backend/flow/requirements.txt
scipy
numpy
seaborn
notebook
pandas
matplotlib
scikit-learn
paho-mqtt